fix: move helper file to inside libs
